음성인식 컴포넌트, 대화요약 컴포넌트, 대화내용보기 페이지로 분리
